libcamera: camera_sensor: Add support for auxiliary image
Some sensor support producing and transmitting an auxiliary image over
a stream separate from the image stream. Add support for this feature in
the CameraSensor interface, and implement it for the CameraSensorRaw
class. The CameraSensorLegacy uses the default stub implementation, as
the corresponding kernel drivers don't support an auxiliary image.

Upstream-Status: Pending [Kernel driver dependency]
Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>