pipeline: neo: Fix pipeline assert terminating streams with isolated IPA
Fatal assert failure below is seen in multi camera operation when IPA is
isolated:

pipeline_handler.cpp:377 assertion "data->queuedRequests_.empty()" failed in stop()

This change revamps the termination of the processing requests in the
pipeline handler to make sure that all requests are correctly completed
in case of stream stop.

The pipeline handler maintains 2 queues for the ongoing requests
submitted by libcamera and not completed yet: the pending queue and
the processing queue.

Requests in the pending queue have not been pushed into the pipeline,
thus have not started the request life cycle starting from the front-end,
then moving to IPA and ISP. Thus, in case of termination, those
requests can have their buffers cancelled and be completed immediately.

Conversely, requests in the processing queue are in flight in the
pipeline. Associated buffers are cancelled and completed along with
their requests. Moreover, the associated Frame::info objects are also
deleted so that those buffers will be ignored during the pipeline
termination.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>