libcamera: ipc_unixsocket: Fix sendSync() timeout and hang
When IPA is running in isolated mode, it is confined in a process
so messages and events are exchanged with the pipeline handler
through a socket-based IPC.

Some issues can be seen with this IPC, typically when multiple
camera are streaming, then stopped simulaneously. This case
triggers the concurrent sending by each pipeline of a synchronous
'close' message to the IPA. The current implementation may show
abnormal delay for sending those messages (timeout) or even hang.

There is a dedicated instance of the IPC socket for each camera,
however pipeline handlers and sockets receiving slots all run in
the same camera manager context.
Current implementation of the synchronous send does blocks after the
sending to wait for the peer response. To do so, sendSync() loops on
dispatchEvent() to let the IPC socket receive slot to be scheduled
to acknowledge the synchronous message.

One issue is that the socket receive slot is received by message by
the blocked thread. That message will resume processEvents() that
can reblock immediately because the slot is not interrupting the
event dispatcher. Subsequent message usually unblocks the event
dispatcher which mitigates the issue, worst case the 2 sec timeout
kicks in and allows the message sending to complete.

The other issue is that in case of the concurrent synchronous
message sending by multiple pipeline handlers, there is a recursion
of sendSync() / processEvents() in the thread stack. As commented
in the source, this is a dangerous construct and it can create hang.
The last synchronous request initiated is the deepest in the stack,
and also the one being executed. It may be delayed because of the
above issue, until its timeout. But its timeout will expire later
than the timeout for preceding messages upper in the stack, whose
sending was initiated before in time. The consequence is that
some timeout are masked by this stack recursion. When unwinding
the stack a pending sendSync() / processEvents() can hang if no other
message is injected in the thread.

This change implements a safer mechanism to handle the synchronous
message sending, using a scheme similar to the one used for non
isolated IPA: the sending thread is blocked until the peer response
is received for that message, without relying on processEvents().
To do that, the IPC socket operation is moved into a dedicated
thread independent of the camera manager. This allows receiving
and processing the socket data ready indication, and unblock the
sender once the synchronous message execution acknowledgement is
available.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>