ipa: neo: camhelper: Switch ox03c10 camhelper to use generic controls
ox03c10 driver has been complemented with the support of the
standard controls V4L2_CID_ANALOGUE_GAIN and V4L2_CID_EXPOSURE,
in addition to the custom low level control.

External IPA is expected to keep usage of custom controls for
now, but the NXP IPA can be switch to usage of the standard
controls.

Thus, introduces a compilation switch to select either the
standard or the custom controls so that the camhelper
implementation can remain in sync between the different IPAs.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>