pipeline: neo: Integrate camera internal endpoints to topology discovery
As camera endpoints support is now available from the CameraSensor class,
use these interfaces to state on the availability of the embedded and
auxiliary streams and their related attributes.

Optional activation of these streams will be configured per camera via
configuration file. This configurability will be implemented in a
separate change.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>