pipeline: neo: Improve colorspace management
This patch aims to rework how the colorspace is set in neo pipeline.
The previous implementation was hard-coding RAW for raw streams and
sRGB  for non-raw streams. Now, it is possible to set different
colorspaces for the Frame stream. Inputs and IR streams are fixed to
RAW. For Frame and IR streams, a try_format call is performed to
validate the colorspace request at the driver level.

A specific case is managing sRGB as libcamera core does not support it
properly. While sRGB is set, the sYCC is enforced just before performing
try/set_format because sRGB conversion is not supported. sYCC will be
converted to V4L2_COLORSPACE_JPEG at the driver level, which is considered
as an abbreviation for V4L2_COLORSPACE_SRGB in v4l2 convention.

Signed-off-by: Alexi Birlinger <alexi.birlinger@nxp.com>