pipeline: neo: Add support for alternated raw stream
The raw stream is currently hard wired to the ISP input0 stream.
There is no stream role defined to designated a possible
auxiliary stream. For sensors exposing an auxiliary stream mapped
into ISP input1, implement an alternated support of the raw
stream where even requests are mapped onto input0 and odd ones
onto input1.
That option is possible only if the two inputs have the same
V4L2 pixel format so that the same buffer can be used for
either inputs.
That strategy may need to be revisited there is a need to support
a camera that expose an auxiliary stream with a format that
differs from the main image stream.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>