pipeline: neo: Add auxiliary and embedded streams to pipeline sequence
With the addition of the auxialiary and embedded optional streams, the
pipeline handler needs to account for the availability of those buffers
when enabled. When present, availability of those buffers is tracked
at NxpNeoFrames::Info level.

To handle the active Request, availability of every active stream buffer
involved (input0 and optionnaly input1 and/or embedded) is awaited
before querying the IPA through fillParamsBuffer.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>