ipa: mojom: neo: Use buffer type map instead of individual variables
Buffers exchanged between the pipeline handler and the IPA are
identified by an identifier (cookie) that are premapped via mapBuffers()
IPC primitive.

Later, pipeline handlers is referring to those buffers in 2 IPA primitives
- fillParamsBuffers()
  To pass the reference to ISP params and raw input0 buffer
- processStatsBuffer()
  To pass the reference to ISP statistics buffer

In the case of HDR, input1 buffer is to be added to fillParamsBuffers().
To handle separate embedded data stream, the embedded buffer id will
also have to be passed to fillParamsBuffers().
Eventually, the frame and ir decoded buffer may have to be passed to
processStatsBuffer(), when available.

To enable support for those additional buffers, replace individual
buffer ids in the fillParamsBuffers() and processStatsBuffer()
arguments by buffer identifier maps.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>