ipa: mojom: neo: Add sensor ID to the init call
The sensor ID is likely to be needed from IPA to be able to select
specific tuning which can differ between sensors of same model.
Hence this patch provides the sensor ID through the init call for the
NEO IPA interface.

However, in case the sensor ID may become part of the IPACameraSensorInfo
as defined from the core mojom, this patch will be reverted.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>