pipeline: neo: Handle queue request error case
When a low level driver returns error while queueing requests, neo
pipeline doesn't propagate this error to pipeline handler, it simply
moves the request to "cancelled" status, but keep it in the pending
queue.

Then at next queue request, the cancelled one is gotten from the
front of the queue. This leads to a crash if low level driver keeps
returning an error. In such case, neo pipeline tries to cancel it
twice, but cancel function verifies the request is in pending state
before cancelling it.

This fix aims at removing the cancelled request from the pending queue
first, and also reporting the error code (if any) to pipeline handler
so it can behave as expected.

Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>