ipa: neo: awb: Clamp gains to prevent divisions by zero
The gain values are currently clamped to the range [0.0, 255.996] used by
the hardware. A zero value makes little sense, as it would completely
remove the contribution of the corresponding color channel from the AWB
accumulators, but worse, would lead to divisions by zero when
calculating the raw means in subsequent iterations. Prevent this by
setting the minimum gain value to 1/256.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>