ipa: neo: Rename "shutter speed" to "exposure time"
The terms "shutter" and "shutter speed" are used through libcamera to
mean "exposure time". This is confusing, both due to "speed" being used
as "time" while it should be the inverse (i.e. a maximum speed should
correspond to the minimum time), and due to "shutter speed" and
"exposure time" being used in different places with the same meaning.

To improve clarity, use "exposure time" consistently to replace
"shutter speed".

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>