ipa: neo: awb: Set default offset values
AWB algorithms applies color channels gain corrections in the
OBWB2 ISP block. In addition to the channel gain, this block
also exposes color channels offset to be used for Black Level
Correction.
This change adds OBWB2 offset default initialization that are
missing, as AWB enables the block. This initialization is added
conditionned to the fact that the block has not been already
enabled, to avoid overriding offsets configured by BLC if it
is executed before.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>