pipeline: neo: Factorize sensor multi-streams implementation
One camera sensor can have up to 3 streams: the long image (ISP input0),
the short image (ISP input1) and embedded data. The long image stream is
mandatory, other are optional depending on the sensor and usage.

Current implementation instanciates (duplicates) code for each stream
in multiple places. This change aims at factorization as possible the
handling of the different streams.

The main changes introduced are:
- neo-utils:
    - CameraInfo::hasStreamXYZ() and CameraInfo::getStreamXYZ() are
      replaced by a more generic CameraInfo::getStream() returning
      an std::optional<stream>
    - CameraInfo::kCameraStreams array is created listing the defined
      camera streams, for stream enumeration purpose.

- neo-pipeline:
    - ISIPipe* NxpNeoCameraData::pipeXYZ_ members removed, replaced by a
      std::map<int, ISIPipe *>, representing optional availability of a
      stream through its pipe
    - V4L2DeviceFormat devFormatXYZ_ members removed, replaced by a
      std::map<int, V4L2DeviceFormat> representing the ISI pipe video
      node for each active camera stream.

A couple of functions could subsequently be shrunk and inlined:
    - NxpNeoCameraData::setupCameraIsiPipes()
    - NxpNeoCameraData::prepareISIPipeBuffers()

A couple of places remain where streams handling is not merged yet. That
is the case for the front end configuration where origin of parameters
varies depending on the stream. This should be aligned once the raw
camera model support is available to the pipeline.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>