ipa: neo: drc: Add dynamic global DRC control
This patch implements 3 modes of operation to determine the
global DRC lookup table:
- the Copy mode: no effective dynamic range compression is performed
and the data passes through as is. This mode is especially
useful for sensors with pixel value bit depth up to 16.
- the Pre-configured mode: the DRC control uses a static predefined
lookup table and a global gain configured in the tuning file.
- the Dynamic mode: the lookup table is generated based on the image
histogram measured by the ISP.

The dynamic mode is based on two principles, namely histogram equalization and
non-linear histogram stretching:
Histogram equalization relies on the linearization of the cumulative distribution
function evaluated from the input histogram - the gain for each bin is calculated
as the ratio between the relative cumulative frequency for a linear histogram
and for the empiric histogram (the real-time luminance histogram measured by the ISP).
Histogram stretching evaluates the empiric dynamic range of the input data,
scales the input histogram according to its real range, and applies a power
function whose exponent represents the strength of the dynamic range compression.

Signed-off-by: Martin Slanina <martin.slanina@nxp.com>