pipeline: neo: Rename embedded-data stream related definitions
Most of the constant and variables related to the embedded-data stream
are been named with a simple 'embedded' radical.
Libcamera code base is mostly using either 'edata' short name or more
explicit 'embedded data' radical for those definitions.
For consistency and conciseness, rename the existing defintion using
one or the other, on a case by case basis.

Associated changes are cosmetic and bring no functional modification.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>