pipeline: neo: Add camera streams definition in config file
Raw camera sensor model exposes optional camera streams that may
be used for embedded data or HDR channels.
Pipeline handler has to know when those channels need to be enabled
when they are supported by a camera. For this purpose, add a cameras
section in the pipeline handler config file, that enables the
configuration of HDR or embedded data streams for a given camera.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>