ipa: neo: Add support for embedded data stream
Current implementation of embedded data in the IPA handles only
the case where data are transmitted as the top lines of the image
data. Add the support for the other case where meta data are
provided as a separate buffer, because they were transmitted by
the camera in a dedicated embedded data stream.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>