ipa: neo: Add padding support for mx95mbcam embedded data
Embedded data format depends on the MIPI-CSI2 Data Type used for
the transmission. In case of mx95mbcam, the byte data stream may
be transported as pixel data padded as 16-bit words, or as 8-bit
values when transported on a dedciated embedded data stream.

This change adds a configuration to the embedded data parser to
support both formats.
In mx95mbcam camHelper, the existing switch ENABLE_EMBEDDED_DATA
is renamed as ENABLE_EMBEDDED_DATA_TOP_LINES to make it clear
that it configures the presence of the embedded data as top lines
of the image. In that case, the embedded data parser is configured
to use 16-bit word format - otherwise 8-bit word format is used as
it is relevant to the embedded data stream.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>