pipeline: neo: Revamp NxpNeoFrames completion and removal
NxpNeoFrames::tryComplete() function is error prone as it may silently
delete the frame Info instance that is passed to it as argument.
To improve readability, update the NxpNeoFrames implementation to
match the one of RkISP1:
- NxpNeoFrames tryComplete() functionis  removed, replaced by a
  tryCompleteRequest() function in the pipeline handler. That function
  explicitly handles the life time of the NxpNeoFrame::Info instance.
- NxpNeoFrames remove() function is replaced by a destroy() function

Along with that change, pipeline handling of the requests and their
associated NxpNeoFrames is complemented with functions:
- tryCompleteRequest() to detect and handle an active request completion
- cancelCompleteRequest() to cancel an active request
Pipeline handler is updated to make usage of those functions.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>