pipeline: neo: Fix queueRequestDevice return value
Function queueRequestDevice() is implemented by each pipeline handler,
invoked by the framework to pass a request from the application.
Return value represents the status of the pipeline with regard to the
queuing of that specific request. In case an error code reported, the
framework cancels and completes the request to notify the application.

In Neo pipeline implementation, this function simply appends the
request to the pending queue. This request will eventually be pushed to
an active request in the processing queue, during one of the subsequent
call to queuePendingRequests(). That function correctly handles the
error handling by notifying the application in case of cancelled request.

Thus, from framework point of view the queueRequestDevice() operation
should always be seen as successfull - error handling if necessary will
be done during the deferred processing of this request within the
queueRequestDevice() function.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>