pipeline: neo: Add configuration for streams buffer count
The number of buffers allocated to a stream is defined in the
StreamConfiguration::bufferCount variable. Pipeline handler
proposes a default value in the GenerateConfiguration() that is
typically not changed by the application.

Current pipeline implementation usess a hardcoded value for this
number of buffers.
This number of buffers has to be big enough to cope with the
latency of the downstream pipeline. But it should also not be
too big because those buffers are expensive in memory usage.

This change adds a section in the platform configuration file
to explicit the number of allocated buffers for the stream.
A new section is created in the file for the configurations
that are global to the pipeline handler. The 'buffer-count' key
is the first one to be implemented.

config.yaml (example)

global:
  buffer-count: 4

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>