neo: pipeline: utils: Remove manual camera stream configuration
The camera streams configuration currently has 2 modes of operation:
- manual definition where streams used by each camera are explicitly
  defined for each entity in the configuration file
- automatic detection where the streams are discovered programatically
  by recursive search in the media device tree

Manual mode of operation was introduced first as a shortcut to use
complex topologies before the automatic mode was ready to do so.
Now that automatic detection is capable of handling all the topologies we
know about so far, remove the implementation for the manual mode that is
obsolete and whose presence is causing confusion.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>