pipeline: neo: Delay multicamera frontend config to acquire() time
Because of the frontend media device limitations, multi-camera use
case requires to preconfigure the cameras graphs before starting any
streaming activity.

That is currently done at match() time. That is a possible issue
because a process may invoke that function to discover the platform
cameras, while the front end media device has been locked by an other
process streaming with one of these cameras.

This change postpones the routing and cameras graphs setup to a later
time when the frontend media device has been locked, meaning that
a first camera has been acquired by the process.
That guarantees that a single process will apply changes to the front
end media device and avoid spurious error messages.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>