pipeline: neo: Remove camera config before controls init
Camera controls are initialized at enumeration and configure() time.
That initialization is currently done from initControls() that will
configure the camera beforehand. This implementation is inherited
from IPU3 pipeline handler but is not necessary for the Neo pipeline.

This change aligns onto the rkisp1 pipeline implementation where the
control list is built by invoking directly updateControls(), thus
bypassing the initControls() function that configures the sensor.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>