pipeline: neo: Aggregate frontend format configuration steps
The frontend media device conguration is currently done in 2 steps,
applicable to each camera stream
1) The graph format configuration from the camera to ISI crossbar
2) The ISI pipe configuration
This requires storage of the format resulting from step 1, to be
later propagated to step 2.

That separation in 2 steps is not necessary, the 2 steps can be
done simultaneously. That simplifies the procedure and remove
the need for this intermediate storage of the stream formats
between the 2 steps.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>