pipeline: neo: Refactor camera configuration selection
Current implementation achieves the camera modes filtering through
usage of the getRawMediaBusFormat() function. It reports only the
sensor sizes relevant to the modes whose format has the highest bit
depth.
Also, multi-camera default configuration management is distributed
into the generateConfiguration() and validate() functions that select
a single size from that prefiltered list.

Purpose of this change is for non multi-camera case to allow the
selection of any of the sizes exposed by the camera driver modes.
The main change is that getRawMediaBusFormat() is replaced
by enumerateRawFormats() that is called only once at camera
enumeration time. That function enquires from the camera its
supported mbus codes and associated sizes then screens the ones
compatible with the pipeline. It produces in turn 2 maps that
are the transposed representation of the same information:
- The list of mbus codes associated to a given size
- The list of sizes associated to a given code
Those 2 maps built at startup are subsequently used by the
generateConfiguration() and validate() function to produce and
filter the StreamConfiguration.
In case of multi-camera condition, those maps are trimmed at
enumeration time to keep the single size and format pair that
associated to the default configuration.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>