ipa: neo: Fix os08a20 BLC offset to the driver-configured value
Os08a20 target BLC offset is configured by the camera driver to a
target value of 64 that is common to all color channels. That is an
absolute offset value applicable to all pixel formats.
For a 12-bit mode, that translates to a 64 * 2^4 = 1024 value after
conversion to the 16-bit format used in the calibration file.

This results in a less contrasted image but with more accurate colors.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>