ipa: neo: blc: Add support for camera bit-depth reference value
In the ISP pipeline, BLC offset is applied after rescaling the camera
pixels to the internal 20-bit depth format.
Thus, the same BLC offset applied by a camera for different pixel
bit-depth formats results in a different value as seen in the 20-bit
ISP internal pipeline. That is because the same input offset has been
rescaled with a different gain.

To be able to apply the proper offset correction, BLC needs to have a
pixel format bit-depth reference for the values it got from the
calibration file, as well as being aware about the current camera mode
bit-depth. With those information it is capable of adjusting the actual
offset to be applied for the current mode.

For that purpose, a reference bit-deth is added to the calibration
file. Also, the actual values to apply are now computed during the
configure() when the camera mode, hence its bit-depth, has been
selected.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>