pipeline: neo: Increase default number of stream buffers
The numbers of streams buffers defines the maximum amount of in-flight
buffers allowed by the pipeline handler. It has to be big enough to cope
with downstream pipeline latency but not too big because it has a direct
impact on the total amount of memory allocated by the pipeline.

Some demanding use case show that current value of 6 too small.
An example of such use case is concurrent streaming of a 8M @30fps
camera with a 2M @60fps camera, with preview display on Wayland.
Latency coming from the ISP decoding time for the 8MP is adding up to
the latency from the display subsystem and may starve the 2M @60fps
stream.

Increase that value to 8 to be more conservative.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>