pipeline: neo: Remove the probed cameras log
Now that a log is output any time that a Neo IPA instance is loaded,
remove the log displaying the number of probed cameras. Its main purpose
was to indicate the user that the Neo pipeline had taken care of those
cameras. It is now redundant with this other log about Neo IPA load
completion.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>