pipeline: neo: Remove name property from config file camera entry
Camera entries may have a name entry in the config file that is not used
for anything else than printing a debug log. It has no functional use
and is redundant with the model or entitity name that have to be
provided.

Remove it for clarity and avoid confusion about its use. The targeted
entity for a camera entry can be documented in the config file using
yaml comment.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>