imx8-isi: Add missing mbus code to the supported YUV/RGB mbus code list
As commented in the code, the sensor mbus code list must be aligned with
ISI driver. Then this patch adds two 2X8 mbus codes which are supported
by the ISI driver.

These mbus codes are exposed by the OV5640 camera driver when using the
DVP (parallel interface) instead of CSI. In order to use a camera
connected to the DVP, these codes must also be added in imx8-isi
pileline's list for proper format enumeration.

For example, the DVP is used as 2nd camera port on i.MX8QXP SoC.

Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>