pipeline: neo: Add camera orientation configuration
A camera configuration is added to the pipeline config file. It may be
used for 2 purposes:
- To specify the camera orientation when application does not support
  the orientation configuration, which is the case with GStreamer
  adapter
- In multi-camera case, to have the flexibility to still configure the
  camera orientation at init time, instead of defaulting to the camera
  mounting orientation.

When specified in the config file, orientation takes precedence over
the configuration passed by the application. If no orientation is
precised in the config file, applications that support it can still
specify the camera orientation.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>