pipeline: neo: Add a test to check ISP driver uAPI version
The ISP driver meta buffers for params and stats do not include a
version number to verify kernel and user space compatibility.
uAPI extensions and versioning will be handled thanks to extensible uAPI
support that is not available yet.

For now, add a basic sanity check to verify the consistency of the
meta buffers allocated and exported by the ISP driver, verifying that
the kernel view of the buffers size matches the one expected by the
user space.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>