ipa: neo: Add Algorithm for ISP PIPE_CONF unit configuration
A subset of the PIPE_CONF unit is now exposed in the uAPI, namely
the INALIGN0/1 and LPALIGN0/1 bitfields of the IMG_CONF_CAM0 register.

An algorithm is added to enable  static configuration for these
values, for the cases where it is necessary to override the default
values applied by the ISP driver.
INALIGN values are not exposed to the calibration file as they are SoC
dependent and intent is to keep calibration file independent from the
hardware.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>