libcamera: camera_sensor: Enable routes to query stream formats
When a V4L2 device supports streams, the route activating a stream
needs to be active so that this stream format can be queried.
For a raw camera sensor, an internal route may be inactive by default
when it is not immutable.

Implementation of CameraSensorRaw::init() probes at startup time the
format of the pads/streams attached to the internal routes of a camera
subdevice, in order to identify the streams it exposes and their type.
To make sure the format of the streams can be retrieved, activate the
internal routes of the subdevice before querying, and restore the
initial route states when finished.

Upstream-Status: Pending [Kernel drivers upstream dependency]
Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>