libcamera: camera_sensor: Add support for auxiliary image
Some sensors can produce and transmit an auxiliary image over a
dedicated stream, separate from the image stream. Examples for such
sensor are the ones supporting HDR or A/B mode captures.

Add support for that type of stream in the CameraSensor interface,
and implement it for the CameraSensorRaw class.
The CameraSensorLegacy uses the default stub implementation, as
the corresponding kernel drivers don't support an auxiliary image.

Upstream-Status: Pending [Kernel driver dependency]
Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>