pipeline: neo: Fix spurious invalid orientation warning
When defining some camera entries in the 'cameras' section of the
pipeline config file, a spurious warning message
"Invalid orientation value 0" may be seen in the console.

That is coming from an incomplete error check about validity of the
orientation configuration. That test is to be done only if the
user has set an orientation value.

Signed-off-by: Julien Vuillaumier <julien.vuillaumier@nxp.com>