ipa: neo: Add LSC algorithm for ISP vignetting unit configuration
This patch provides the Lens Shading Correction algorithm which applies
multipliers to all pixels to compensate for the lens shading effect.
The YAML tuning file is used to define:
- each coefficient specified with a set of LUT composed for each color channel.
- the block count used to partition the image.

Those parameters are dependent on the sensor resolution and are provided
for each supported sensor resolution.
If one sensor resolution has no LSC parameters defined, the LSC computation
is disabled for this resolution.

Furthermore, each LUT is defined for a specific color temperature.
The algorithm interpolates the LUT according to the measured color temperature.
Hence the LSC algorithm depends on the AWB algorithm which is measuring
the color temperature. For this reason, the LSC algorithm should run after the
AWB algorithm.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>