ipa: neo: Enable vignetting configuration for ox03c10
For each of the sensor profile (resolution), the LSC algorithm
is defined with:
- "block-count": Number of blocks used to partition the image
- "sets": Set of LUT entries composed for r/g/b channels and
for a specific color temperature.

The ox03c10 tuning is for 3 illuminant modes:
- Halo: <2800K,
- Fluorescent (indoor/LED): [3600K-4200K]
- D65 (daylight): >5000K
Interpolations are between 2800K-3600K and between 4200K-5000K.
Hence LUTs are the same:
- for ct=0K and ct=2800K
- for ct=3600K and ct=4200K

The values come from the MMS tuning of the SW7 MS48 release.

Signed-off-by: Celine Laurencin <celine.laurencin@nxp.com>